requirements.txt
fake-useragent==0.1.11
python-dotenv==0.19.0
selenium==3.141.0
urllib3==1.26.6
